+++ src/ui/components/textinput.cpp
@@ -17,7 +17,7 @@ ui::TextInput::TextInput(
         ServiceProvider* service_provider,
         Font font,
         const Color& color,
-        usize focus_id,
+        u32 focus_id,
         const shapes::URect& fill_rect,
         const Layout& layout,
         bool is_top_level
@@ -47,7 +47,7 @@ ui::TextInput::TextInput(
         ServiceProvider* service_provider,
         Font font,
         const Color& color,
-        usize focus_id,
+        u32 focus_id,
         std::pair<double, double> size,
         Alignment alignment,
         const Layout& layout,
